Reinforcement Learning- OpenAI Cartpole
This project explores the application of reinforcement learning (RL) techniques, specifically Q-learning and Double SARSA, to solve the CartPole problem. The CartPole problem is a classic RL challenge where the goal is to balance a pole on a cart. The project aims to compare the performance of these two RL models, highlighting their strengths and weaknesses in achieving stable and efficient solutions.